Refactor : client/config의 반응형 뷰와 관련된 상수들을 분리
상세 :
- client/constants/responsiveView에서 반응형 뷰 관련 상수들 관리

#388
