+++ client/src/hooks/useShiftingToWhichView.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import useIsMobile from './useIsMobile';
-import { MOBILE_VIEW, DESKTOP_VIEW } from '../config';
+import { MOBILE_VIEW, DESKTOP_VIEW } from '../constants/responsiveView';
 
 const useShiftingToWhichView = mobileViewBreakpoint => {
   const isMobile = useIsMobile(mobileViewBreakpoint);
